+++ packages/cortex-react/src/components/DatePicker/DateRangePickerInput.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useDateRangePickerInput } from '../../hooks';
+import { dateToCalendarDate } from '../../utils';
 import { DateRange, RangeCalendar } from '../Calendar/RangeCalendar';
 import { InputPropsBase } from '../Input';
-import { dateToCalendarDate } from '../utils';
-import { DateField } from './DateField';
-import { DatePickerInputBase } from './DatePickerInputBase';
 import { Popover } from '../Popover';
 import { usePopoverContext } from '../Popover/Context';
 import { Content } from './Content';
+import { DateField } from './DateField';
+import { DatePickerInputBase } from './DatePickerInputBase';
 
 export interface DateRangePickerInputProps extends InputPropsBase {
   value?: DateRange;