+++ packages/cortex-react/src/utils/date.ts
@@ -1,14 +1,4 @@
 import { CalendarDate, getLocalTimeZone } from '@internationalized/date';
-import React from 'react';
-
-export const getNameInitials = (name: string) => {
-  const nameSplit = name.split(' ');
-  const length = nameSplit.length;
-  if (length > 1) {
-    return `${nameSplit[0][0]}${nameSplit[length - 1][0]}`;
-  }
-  return name[0];
-};
 
 export const dateToCalendarDate = (value?: Date) => {
   return new CalendarDate(
@@ -21,10 +11,3 @@ export const dateToCalendarDate = (value?: Date) => {
 export const calendarDateToDate = (value: CalendarDate) => {
   return value.toDate(getLocalTimeZone());
 };
-
-export function cloneWithProps(
-  children: React.ReactElement,
-  triggerProps: any
-) {
-  return React.cloneElement(<div>{children}</div>, triggerProps);
-}