+++ packages/cortex-react/src/utils/react.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function cloneWithProps(children: React.ReactElement, props: any) {
+  return React.cloneElement(<div>{children}</div>, props);
+}